Module aws_sdk_braket::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

device_offline_exception

See DeviceOfflineException

device_retired_exception

See DeviceRetiredException

internal_service_exception

See InternalServiceException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

CancelQuantumTaskError

ConflictException

An error occurred due to a conflict.

CreateQuantumTaskError

DeviceOfflineException

The specified device is currently offline.

DeviceRetiredException

The specified device has been retired.

GetDeviceError

GetQuantumTaskError

InternalServiceException

The request processing has failed because of an unknown error, exception, or failure.

ListTagsForResourceError

ResourceNotFoundException

The specified resource was not found.

SearchDevicesError

SearchQuantumTasksError

ServiceQuotaExceededException

The request failed because a service quota is exceeded.

TagResourceError

ThrottlingException

The throttling rate limit is met.

UntagResourceError

ValidationException

The input fails to satisfy the constraints specified by an AWS service.

Enums

CancelQuantumTaskErrorKind

CreateQuantumTaskErrorKind

GetDeviceErrorKind

GetQuantumTaskErrorKind

ListTagsForResourceErrorKind

SearchDevicesErrorKind

SearchQuantumTasksErrorKind

TagResourceErrorKind

UntagResourceErrorKind